modificacion de menú y estilos + responsive, faltan resposiive de tab…

+++ src/Components/TableUsers/TableUsers.js
@@ -5,9 +5,11 @@ import './TableUsers.css'
 const TableUsers = () => {
    
 	const deleteUser = async (_id) => {
-		await axios.delete(`https://burguercode-db.onrender.com/deleteuser/${_id}`)
-		alert('Usuario eliminado correctamente')
-		window.location.href = "/users"
+		if (window.confirm('¿Estás seguro que deseás eliminar este usuario?')) {
+			await axios.delete(`https://burguercode-db.onrender.com/deleteuser/${_id}`)
+			alert('Usuario eliminado correctamente')
+			window.location.href = "/adminusers"
+		}
 	}
   
   const [user, setUser] = useState([]);
@@ -50,9 +52,9 @@ const TableUsers = () => {
 	<thead>
 	<tr>
 		<th>NOMBRE</th>
-		<th>EMAIL</th>
-		<th>DIRECCIÓN</th>
-		<th>TELÉFONO</th>
+		<th className="movile-users-css">EMAIL</th>
+		<th className="movile-users-css">DIRECCIÓN</th>
+		<th className="w-auto text-center">TELÉFONO</th>
 		<th></th>
 
 	</tr>
@@ -61,15 +63,17 @@ const TableUsers = () => {
 	{usersOrganized.map(user=>
 	<tr className="viewunique">
 		 <td><b>{user.lastname} {user.name}</b></td>
-		 <td>{user.email}</td>
-		 <td><i>{user.address}</i></td>
-		 <td>{user.phone}</td>
+		 <td className="movile-users-css">{user.email}</td>
+		 <td className="movile-users-css"><i>{user.address}</i></td>
+		 <td className="w-auto text-center">{user.phone}</td>
      <td>
-			<button className="btn btn-danger text-center d-flex justify-content-center p-2 m-1 rounded-0 border-0" onClick={(_id) => deleteUser(user._id)} type="button" key={user._id}>
+			<div className="d-flex justify-content-center">
+			<button className="btn btn-danger p-2 m-1 rounded-0 border-0" onClick={(_id) => deleteUser(user._id)} type="button" key={user._id}>
 			<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16" alt="Eliminar usuario">
   <path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z"/>
 </svg>
 			</button>
+			</div>
 		 </td>
 	</tr>
 	)}
